fix: scientific notation in usePrices
