+++ packages/react-query/src/hooks/prices/usePrices.ts
@@ -1,5 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 import ms from 'ms'
+import { withoutScientificNotation } from 'sushi/format'
 import { Fraction } from 'sushi/math'
 import { getAddress, isAddress, parseUnits } from 'viem'
 
@@ -16,10 +17,11 @@ export const usePrices = ({ chainId }: UsePrices) => {
         // `http://localhost:3001/v2/${chainId}`,
       ).then((response) => response.json())
       return Object.entries(data).reduce<Record<string, Fraction>>(
-        (acc, [address, price]) => {
-          if (isAddress(address)) {
+        (acc, [address, _price]) => {
+          const price = withoutScientificNotation(_price.toFixed(18))
+          if (isAddress(address) && typeof price !== 'undefined') {
             acc[getAddress(address)] = new Fraction(
-              parseUnits(price.toFixed(18), 18).toString(),
+              parseUnits(price, 18).toString(),
               parseUnits('1', 18).toString(),
             )
           }
